virtapt: do not require /etc/apt/trusted.gpg
Before elbe adds any keys specified in the xml file, setup_gpg()
copies /etc/apt/trusted.gpg from the host to use as the elbe apt
keyring. This was implemented in b78d59f8114f ("virtapt: use
etc/apt/trusted.gpg instead of etc/apt/trusted.gpg.d"). However,
trusted.gpg only contains locally added keys and may not exist.

Only copy trusted.gpg from the host if it actually exists. If it
does not exist, setup_gpg() will create an empty keyring for use
elbe to use.

+++ elbepack/virtapt.py
@@ -162,7 +162,8 @@ class VirtApt:
                   "fix this problem")
             sys.exit(20)
 
-        system('cp /etc/apt/trusted.gpg "%s"' % ring_path)
+        if os.path.exists("/etc/apt/trusted.gpg"):
+            system('cp /etc/apt/trusted.gpg "%s"' % ring_path)
 
         gpg_options = '--keyring "%s" --no-auto-check-trustdb ' \
                       '--trust-model always --no-default-keyring ' \
